v2.1 update for KTX-Software and Khronos file format KTX2 compatibility. This update impacts the UASTC HDR 4x4, UASTC HDR 6x6i, and XUASTC LDR formats. New KTX2 files in these specific formats using v2.1 can't be loaded using v2.0, however we've kept backwards compatibility with all of our previously written KTX2 files. (v2.1 can transcode all previously written files.) This is part of our effort to be compatible with Khronos's specification. bc6hf encoder's smooth block path updated to handle inputs with texels close to MAX_HALF_FLOAT. Adding command line options to basisu tool so it can write either v2.0 or v1.6 compatible UASTC HDR 6x6i files. The default is v1.6. Updating WebGL KTX2 testbed so on ASTC devices it automatically switches between sRGB or linear internal formats for the created texture.

+++ example_capi/example_capi.c
@@ -88,7 +88,7 @@ int write_tga_image(const char* pFilename, int w, int h, int has_alpha, const ui
         return -6; // overflow bogus dimensions
 
     /* allocate one scanline for BGRA/BGR output */
-    size_t row_bytes = (size_t)w * bytes_per_pixel;
+    size_t row_bytes = (size_t)((size_t)w * bytes_per_pixel);
     uint8_t* pRow = (uint8_t*)malloc(row_bytes);
     if (!pRow) 
     {
@@ -539,11 +539,11 @@ int test_2D()
     }
 
     // Write the data to disk
-    write_blob_to_file("test.ktx2", pComp_data, comp_size);
+    write_blob_to_file("test.ktx2", pComp_data, (size_t)comp_size);
     printf("Wrote file test.ktx2\n");
 
     // Now inspect and transcode the .KTX2 data to png/astc files
-    if (!transcode_ktx2_file(pComp_data, comp_size, "2D"))
+    if (!transcode_ktx2_file(pComp_data, (size_t)comp_size, "2D"))
     {
         fprintf(stderr, "transcode_ktx2_file() failed\n");
         return EXIT_FAILURE;
@@ -646,11 +646,11 @@ int test_2D_array(BOOL tex_video_flag, int L, BOOL mipmap_flag)
 
     // Write the data to disk
     snprintf(filename_buf, sizeof(filename_buf), "test_%s.ktx2", pDesc);
-    write_blob_to_file(filename_buf, pComp_data, comp_size);
+    write_blob_to_file(filename_buf, pComp_data, (size_t)comp_size);
     printf("Wrote file %s\n", filename_buf);
 
     // Now inspect and transcode the .KTX2 data to png/astc files
-    if (!transcode_ktx2_file(pComp_data, comp_size, pDesc))
+    if (!transcode_ktx2_file(pComp_data, (size_t)comp_size, pDesc))
     {
         fprintf(stderr, "transcode_ktx2_file() failed\n");
         return EXIT_FAILURE;
